feat(server): moved achievement unlock evaluation inside the evaluator
- feat(server): changed the AchievementEvaluator methode to return a boolean for now
- represent if the Achievement was unlocked during the progress update
- for now a boolean return is sufficient but can be a more complex EvaluatorResult in the future

+++ server/application-server/src/main/java/de/tum/in/www1/hephaestus/achievement/UserAchievement.java
@@ -112,7 +112,7 @@ public class UserAchievement {
      */
     @Column(name = "current_value", nullable = false)
     @Builder.Default
-    private long currentValue = 0;
+    private int currentValue = 0;
 
     @PrePersist
     protected void onCreate() {
